src/services/backend/requirements.txt
asgiref==3.4.1
Django==4.0
django-cors-headers==3.10.1
djangorestframework==3.12.4
gunicorn==20.1.0
Pillow==8.4.0
sqlparse==0.4.2
tzdata==2021.5
whitenoise==5.3.0
